requirements.txt
torch
torchvision
numpy
opencv-python
Pillow
matplotlib
scipy
scikit-learn
torchsummary
PyYAML
tqdm
mlflow
fastapi
uvicorn
prometheus-client
pytest
requests
